refactor : formatAuthors ,formatNaruApiBookDataToBookItemData 리팩토링 및 폴더 위치 변경
