+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './date';
 export * from './debounce';
 export * from './xmlParser';
 export * from './errorHandler';
+export * from './bookDataFormatter';
